chore(condo) DOMA-6175: reset users and organizations after test run
